Добавление 5 пункта дублирования файла и 6 удаление дубликатов

+++ les01/les04.py
@@ -15,7 +15,7 @@ print(name, ", добро пожаловать в мир Python!")
 
 answer = ''
 # PEP-8
-while answer != 'q' or 'Q':
+while answer != 'q':
     answer = input("Давай поработаем? (Y/N/Q)")
     if answer == 'Y' or answer == 'y':
         print("Чем бы ты хотел заняться?")
@@ -23,7 +23,9 @@ while answer != 'q' or 'Q':
         print("[1] - вывести список файлов")
         print("[2] - выведу информацию о системе")
         print("[3] - выведу список процессов")
-        print("[4] - дублирование файлов")
+        print("[4] - продублировать файлы в текущей директории")
+        print("[5] - продублировать указаный файл")
+        print("[6] - удалить дубликаты файлов")
         do = int(input("Укажите номер действия: "))
 
         if do == 1:
@@ -42,8 +44,25 @@ while answer != 'q' or 'Q':
             file_list = os.listdir()
             i = 0
             while i < len(file_list):
-                newfile = file_list[i] + '.dupl'
-                shutil.copy(file_list[i], newfile) #Копируй
+                if os.path.isfile(file_list[i]):
+                    newfile = file_list[i] + '.dupl'
+                    shutil.copy(file_list[i], newfile) #Копируй
+                i += 1
+        elif do == 5:
+            print("=Дублирование указанного файла=")
+            filename = input("Укажите название файла: ")
+            if os.path.isfile(filename):
+                newfile = filename + '.dupl'
+                shutil.copy(filename, newfile) #Копируй
+        elif do == 6:
+            print("=Удаление дубликатов в директории=")
+            dirname = input("Укажите имя директории: ")
+            file_list = os.listdir(dirname)
+            i = 0
+            while i < len(file_list):
+                fullname = os.path.join(dirname, file_list[i])  #\ /
+                if fullname.endswith('.dupl'):
+                    os.remove(fullname)
                 i += 1
         else:
             input("Ты придумал что то ещё?")
